move price impact to a new component
