+++ src/components/organisms/AssetActions/Trade/Slippage.tsx
@@ -1,14 +1,10 @@
 import { FormikContextType, useFormikContext } from 'formik'
-import React, { ChangeEvent, ReactElement, useEffect, useState } from 'react'
+import React, { ChangeEvent, ReactElement } from 'react'
 import { FormTradeData, slippagePresets } from '../../../../models/FormTrade'
 import InputElement from '../../../atoms/Input/InputElement'
 import styles from './Slippage.module.css'
 
-export default function Slippage({
-  priceImpact
-}: {
-  priceImpact: string
-}): ReactElement {
+export default function Slippage(): ReactElement {
   // Connect with form
   const { setFieldValue, values }: FormikContextType<FormTradeData> =
     useFormikContext()
@@ -18,24 +14,18 @@ export default function Slippage({
   }
 
   return (
-    <>
-      <div className={styles.slippage}>
-        <div>
-          <strong>Expected price impact</strong>
-          <strong>{` ${priceImpact}%`}</strong>
-        </div>
-        <strong>Slippage Tolerance</strong>
-        <InputElement
-          name="slippage"
-          type="select"
-          size="mini"
-          postfix="%"
-          sortOptions={false}
-          options={slippagePresets}
-          value={values.slippage}
-          onChange={handleChange}
-        />
-      </div>
-    </>
+    <div className={styles.slippage}>
+      <strong>Slippage Tolerance</strong>
+      <InputElement
+        name="slippage"
+        type="select"
+        size="mini"
+        postfix="%"
+        sortOptions={false}
+        options={slippagePresets}
+        value={values.slippage}
+        onChange={handleChange}
+      />
+    </div>
   )
 }
